pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pdf2anki"
version = "0.1.0"
description = "Convert PDF documents to Anki flashcards using LLMs"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Nicholas McCarthy", email = "nicholas@example.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Education",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
keywords = ["anki", "pdf", "flashcards", "llm", "education"]
requires-python = ">=3.9"
dependencies = [
    "typer[all]>=0.12.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "PyYAML>=6.0",
    "jinja2>=3.1.0",
    "PyMuPDF>=1.23.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.8",
    "openai>=1.0.0",
    "genanki>=0.13.0",
    "rich>=13.0.0",
    "tiktoken>=0.5.0",
    "python-dotenv>=1.0.0",
    "rapidfuzz>=3.0.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "hashlib_compat>=1.0.0",
]

[project.optional-dependencies]
ocr = [
    "pytesseract>=0.3.10",
    "Pillow>=10.0.0",
]
tables = [
    "camelot-py[cv]>=0.10.0",
    "tabula-py>=2.8.0",
]
rag = [
    "faiss-cpu>=1.7.4",
    "chromadb>=0.4.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "pre-commit>=3.0.0",
]
all = [
    "pdf2anki[ocr,tables,rag]",
]

[project.scripts]
pdf2anki = "pdf2anki.cli:app"

[project.urls]
Homepage = "https://github.com/NicholasMcCarthy/pdf2anki"
Repository = "https://github.com/NicholasMcCarthy/pdf2anki"
Issues = "https://github.com/NicholasMcCarthy/pdf2anki/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["pdf2anki"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "genanki.*",
    "fitz.*",
    "tiktoken.*",
    "faiss.*",
    "chromadb.*",
    "camelot.*",
    "tabula.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=src/pdf2anki",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-fail-under=80",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]